Submit button works; graphics update
To colin: the submit button now functions properly. It runs through 3
times, and goes to result screen after the 3rd submit.  To sergio: the
cup sizes are text objects, not actual values. I just added them in for
demo purposes. I'll fix that so a running total can be tracked, but I
just wanted to get this out for you guys asap so you can continue
working on it.

Also added splash page image.