Fractions integrated in problem validation
The Fraction class is integrated in the project and the problem screen
evaluates by value rather than number of clicks. The problem is still
hardcoded, but can be adapted rather easily.
Also I'm experiencing significant slowdown when loading the problem
screen, but I'm not sure if this is due to the Fraction class or the low
RAM on the lab computers. Can anyone verify this?