The construct is there
I got frustrated spending too much time trying to get a simple process
done of piping the object info from order_list to problem_screen. I've
commented out where I was modfying code to try to make it work. i
resolved null pointer errors for this commit. When I committ it says
"deleted" next to some; not sure if that is on my end, or from someone
else. If this is wrong, roll back the commit.