Slight problem screen layout modifications
So now there are 6 placeholder layouts in the problem screen. I changed
this so there can be text, and with text I can eliminate the ImageView
and just use the drawable field in the TextView. There is also a submit
button but it does nothing.
I also have not yet eliminated the customer screen as we discussed in
the meeting.